requirements-dev.txt
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Linting
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Type checking
types-setuptools>=67.0.0
pandas-stubs>=1.5.3

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0

# Development
pre-commit>=3.0.0
tox>=4.0.0
